refactor(vrops): Moved vrops store models to the correct place Signed-off-by: Stefan Genov <stefan.genov@broadcom.com>
