+++ common/artifact-manager/src/main/java/com/vmware/pscoe/iac/artifact/aria/operations/store/VropsPackageStore.java
@@ -70,8 +70,8 @@ import com.vmware.pscoe.iac.artifact.model.Package;
 import com.vmware.pscoe.iac.artifact.model.PackageContent;
 import com.vmware.pscoe.iac.artifact.model.PackageContent.Content;
 import com.vmware.pscoe.iac.artifact.model.Version;
-import com.vmware.pscoe.iac.artifact.model.vrops.VropsPackageDescriptor;
-import com.vmware.pscoe.iac.artifact.model.vrops.VropsPackageMemberType;
+import com.vmware.pscoe.iac.artifact.aria.operations.store.models.VropsPackageDescriptor;
+import com.vmware.pscoe.iac.artifact.aria.operations.store.models.VropsPackageMemberType;
 import com.vmware.pscoe.iac.artifact.aria.operations.rest.RestClientVrops;
 import com.vmware.pscoe.iac.artifact.aria.operations.models.AlertDefinitionDTO;
 import com.vmware.pscoe.iac.artifact.aria.operations.models.AuthGroupDTO;
