+++ common/artifact-manager/src/main/java/com/vmware/pscoe/iac/artifact/aria/operations/store/models/VropsPackageMemberType.java
@@ -12,7 +12,7 @@
  * This product may include a number of subcomponents with separate copyright notices and license terms. Your use of these subcomponents is subject to the terms and conditions of the subcomponent's license, as noted in the LICENSE file.
  * #L%
  */
-package com.vmware.pscoe.iac.artifact.model.vrops;
+package com.vmware.pscoe.iac.artifact.aria.operations.store.models;
 
 /**
  * VropsPackageMemberType defines the asset types that vROPs package supports.
