+++ common/artifact-manager/src/test/java/com/vmware/pscoe/iac/artifact/aria/operations/store/VropsPackageStoreTest.java
@@ -38,8 +38,8 @@ import com.vmware.pscoe.iac.artifact.cli.CliManagerVrops;
 import com.vmware.pscoe.iac.artifact.model.Package;
 import com.vmware.pscoe.iac.artifact.model.PackageFactory;
 import com.vmware.pscoe.iac.artifact.model.PackageType;
-import com.vmware.pscoe.iac.artifact.model.vrops.VropsPackageDescriptor;
-import com.vmware.pscoe.iac.artifact.model.vrops.VropsPackageMemberType;
+import com.vmware.pscoe.iac.artifact.aria.operations.store.models.VropsPackageDescriptor;
+import com.vmware.pscoe.iac.artifact.aria.operations.store.models.VropsPackageMemberType;
 import com.vmware.pscoe.iac.artifact.aria.operations.rest.RestClientVrops;
 import com.vmware.pscoe.iac.artifact.aria.operations.models.AuthGroupDTO;
 import com.vmware.pscoe.iac.artifact.aria.operations.models.AuthUserDTO;
