+++ maven/plugins/vrops/src/main/java/com/vmware/pscoe/maven/plugins/PackageMojo.java
@@ -39,7 +39,7 @@ import com.vmware.pscoe.iac.artifact.configuration.ConfigurationException;
 import com.vmware.pscoe.iac.artifact.model.Package;
 import com.vmware.pscoe.iac.artifact.model.PackageFactory;
 import com.vmware.pscoe.iac.artifact.model.PackageType;
-import com.vmware.pscoe.iac.artifact.model.vrops.VropsPackageDescriptor;
+import com.vmware.pscoe.iac.artifact.aria.operations.store.models.VropsPackageDescriptor;
 
 import edu.emory.mathcs.backport.java.util.Arrays;
